avoid using java.File.renameTo as it is unreliable
In testing on Ubuntu this method always fails (for me at least) to download the file; there are no exceptions but the specified destinationFile is never created. Debugging the issue showed that the call to renameTo always returned false. I can't see any good reason why the rename should be failing in my setup but given that the method is notoriously unreliable I've switched to using Files.move(src,dest) from the Google Common IO library. This appears to work 100% of the time in the testing I've done so far.